Personal Code: Logic for Keep Track Show All page
My biggest achievement with this block of code was with the sorting. Sorting (and one level of sub-sorting) on related tables, for me, was tricky. But I managed to figure out the sorting callback for 2 different scenarios, particularly the blocks at lines 167 and 183. It took some days to figure this out... advanced documentation was non-existent in Laravel, but the Laravel (sort) code, the PHP docs, and Google/Stack Overflow helped a lot.